+++ packages/stream_chat_flutter/lib/src/autocomplete/stream_mention_autocomplete_options.dart
@@ -242,6 +242,7 @@ class _StreamMentionAutocompleteOptionsState extends State<StreamMentionAutocomp
   }
 
   Future<List<User>> _fetchUsers(String query) async {
+    if (!widget.channel.canCreateMention) return const [];
     try {
       if (widget.mentionAllAppUsers) {
         return await _queryUsers(query);